linear/addrge_api.h: Restore to avoid ci issue
